Adding github static site generator
The generator is not very good, and the design is fairly bad.
It's more or less a placeholder for something when more/better docs
is needed, if the project ever gathers users.

The script will generate content from the EDoc documentation, apply the
styles from site/screen.css and make something nicer than the default
OTP frame-based layout.