Add snapshot functionality for offline inspections
When trying to debug memory issues it is not always possible to do
so on the target system. Snapshots in recon_alloc allows the user to
easily save a memory snapshot on the target node and then later analyze
it on some other machine.

The names of the functions in the API (i.e. snapshot_Action) was chosen in
order for the functions to be grouped in the documentation and not clutter
it.

See the documentation in snapshot_load/1 for an example use-case.