Add unit conversion for easier comparisons
When working with memory analysis, number often are quite big and become hard
to compare. Therefore it is nice to be able to work in giga/mega/kilo-bytes
depending on what memory you are currently looking at.

This commit adds support for changing which unit is used by all functions that
return a memory size.