Change average_sizes to average_block_sizes
When figuring out what sizes to have in your lmbcs, smbcs or sbct it
is the average block size that is interesting. This is because we
usually want to modify the limits to fit as many blocks as we want in
one carrier. The average carrier size is easy enough to figure out by
just looking at the configurations of the allocators.

This is a backwards incompatible change.