Allow multiple patterns to be traceable
This reworks the API by adding more options -- passing in tuples and/or
lists of tuples representing the MFAs available in calls/4 and calls/5,
and adding a more generic calls/2.

Internally, the order is reversed on a few calls so that all patterns
are added at once, and then processes enabled, allowing a once-at-a-time
trigger of all submitted patterns.