Add stricter matching for inexisting allocators
This would cause bugs in R15B03 when loading crash dumps from more
modern commands used in R16 and others to generate dumps in instances
where a given allocator isn't explicitly present.

In this case the value returned is {Alloc, false} for that allocator
rather than {Alloc, ListOfAllocs}, which caused the multi-generator LC
to crash.

This one adds an assertion as a list on a generator to make sure to
filter out empty allocators.