Broadening recon_trace:calls/2 to allow >1 tspec
Only calls/3 allowed for more than one trace specification to be used,
creating a somewhat inconsistent API between with and without options.

This patch makes sure that multiple specs are allowed in both calls/2
and calls/3 for a more consistent interface, and is backwards compatible
(a superset) of the currently existing interface.