Include mbcs_pool in all mbcs values
In R16B02 an option to enable a mbc pool was added. In order to
get the total size of current mbcs we have to include the pool in
all our calculations with mbcs.

This commit also does a bit of refactoring that makes the code
more modular when fetching info from within the allocator statistics.