unexpected early strings binary/code
ran into errors when "binary" and "code" strings appeared before the expected count. resulting in an error when it tried to /(1024*1024), found that ensuring we use ^ to define start of word solved this. I updated all of the proposed strings to account for a chance that this occurs for others.

./erl_crashdump_analyzer.sh: line 27: let: m=0x00007f4da25e6358:SReturn addr 0xACE80DB8 (*redacted*_binary:to_utf8/1 + 128)/(1024*1024): syntax error in expression (error token is ":SReturn addr 0xACE80DB8 (*redacted*_binary:to_utf8/1 + 128)/(1024*1024)")