Optimize (not) printing empty string from formatter
Sometimes it is handy to implement some additional filtering in a
custom formatter function. In this case the formatter can return the
empty string if it wants to ignore a trace message. Optimize the
formatting so in the empty string case it does not send an io request
to the IO server. This avoids a "flickering prompt" effect (when the
empty string is printed) in case of 1000s of ignored trace messages
per second.